+++ packages/toolkit/src/view/pipeline-builder/nodes/CustomNode.tsx
@@ -210,7 +210,7 @@ const NameRow = (props: { name: string; definition: ConnectorDefinition }) => {
       );
       break;
     }
-    case "CONNECTOR_TYPE_SOURCE": {
+    case "CONNECTOR_TYPE_OPERATOR": {
       fallbackImage = (
         <Icons.Box className="h-4 w-4 stroke-semantic-fg-primary" />
       );